pyproject.toml
[tool.poetry]
name = "ptx_classification"
version = "0.1.0"
description = ""
authors = ["Charlotte Gerhaher"]

[tool.poetry.dependencies]
python = "^3.8"
ciip-dataloader = {path = "../ciip-dataloader/", develop = true}
numpy = "^1.23.0"
matplotlib = "^3.5.2"
scikit-learn = "^1.1.1"
pandas = "^1.4.3"
bokeh = "^2.4.3"
torch = "^1.12.0"
pytorch-lightning = "^1.6.4"
ray = {extras = ["tune"], version = "^1.13.0"}
Pillow = "^9.2.0"
seaborn = "^0.11.2"
torchvision = "^0.13.0"
pydicom = "^2.3.0"
grad-cam = "^1.4.0"
openpyxl = "^3.0.10"
psutil = "^5.9.1"
chord = "^1.0.2"
holoviews = "^1.15.0"
plotapi = "^5.1.0"


[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.3.0"
pytest = "^7.1.1"
mypy = "^0.961"

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
